requirements.txt
# PyTorch CPU wheels (ADD this index, don't replace PyPI)
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.3.1
torchvision==0.18.1

numpy==1.26.1
Pillow
huggingface_hub
einops
safetensors

rerun-sdk=0.25.1
gradio
matplotlib

